criacao de uma classe dedicada para os testes de contrato
