+++ src/funcionaltest/java/br/com/covid/brazil/api/funcional/ObterDadosCovidFuncionalTest.java
@@ -11,9 +11,7 @@ import static br.com.covid.brazil.api.util.MatchersConstantes.assertBodyDefaultD
 import static br.com.covid.brazil.api.util.MatchersConstantes.assertBodyDefaultDataPorPosicao;
 import static br.com.covid.brazil.api.util.RotasEnum.*;
 import static br.com.covid.brazil.api.util.TesteConstantes.*;
-import static io.restassured.module.jsv.JsonSchemaValidator.matchesJsonSchemaInClasspath;
 import static java.lang.String.format;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -23,23 +21,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @Sql(value = "classpath:sql/InserirDadosFuncional.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 @Sql(value = "classpath:sql/DeletarDadosFuncional.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
 class ObterDadosCovidFuncionalTest extends BaseTestFuncional {
-
-    @Test
-    @Tag("funcional")
-    @DisplayName("Deve Validar Contrato - Obter Dados Covid Brasil.IO")
-    void deveRetornarSucessoAoValidarContratoObterDadosBrasilIo() throws Exception {
-        String response =
-                getMvc().perform(get(OBTER_DADOS_COVID_BRASIL_IO.getUrl())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .param(UF_PARAM_BODY, RS_PARAM)
-                                .param(MUNICIPIO_PARAM_BODY, ALEGRE_PARAM))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assertThat(response, matchesJsonSchemaInClasspath("schemas/obterDadosCovidBrasilIo_schema.json"));
-    }
 
     @Test
     @Tag("funcional")
